Load and save game
Added a UI on the top left of the screen.
Under the heihercy i created two buttons. Save and load game.
Save game will store the player position on the map and the scene he was currently in to a hidden location on the computer.

Load Game: reads from the saved game txt file. and restores the character information stored.

Added 2 new scripts and updated alice controller script.

Player data is a data storing object script.

Save system script has the main functionality of save and load game

Added load and save functions to the alice script